Fix shutdown
Currently if the script gets SIGINT/SIGTERM while it's sleeping, the sleep is
not interrupted. That means it takes 15 minutes to shut down.

The fix is to use a threading.Event - instead of time.sleep we can call its wait
method with a timeout. This returns True if the event has been set, indicating a
signal was received.